refactor(agentic): modularize post matching and dispatch helpers (#93, #94)
