+++ backend/app/services/agentic/verification_graph.py
@@ -15,8 +15,7 @@ from app.services.agentic.schemas import (
 )
 from app.services.agentic.tools.common import resolve_session
 from app.services.agentic.verification_matching import (
-    format_canonical_post_url,
-    match_post_on_timeline,
+    evaluate_target_post,
     probe_url_reachability,
 )
 from app.services.agentic.verification_scraper import scrape_x_profile_feed
@@ -154,71 +153,6 @@ async def scrape_x_profile_timeline_node(
     return {"timeline_tweets": timeline_tweets, "status": "timeline_scraped"}
 
 
-def _verify_single_x_post(
-    *, post: dict[str, Any], timeline_tweets: list[dict[str, Any]]
-) -> VerificationItemReport:
-    """Perform ground-truth verification on an X post."""
-    post_id = str(post.get("id") or "")
-    content = str(post.get("content") or "")
-    ext_id = post.get("external_post_id")
-
-    is_match, matched_text, matched_id, conf = match_post_on_timeline(
-        expected_content=content,
-        expected_ext_id=ext_id,
-        timeline_tweets=timeline_tweets,
-    )
-
-    live_url = format_canonical_post_url(platform="x", ext_id=matched_id or ext_id)
-
-    return VerificationItemReport(
-        post_id=post_id,
-        platform="x",
-        is_verified=is_match,
-        external_post_id=matched_id or ext_id,
-        matched_text=matched_text,
-        match_confidence=conf,
-        live_url=live_url,
-        status_code=200 if is_match else 404,
-        error=None if is_match else "Post not found on live profile timeline",
-    )
-
-
-def _verify_single_linkedin_post(*, post: dict[str, Any]) -> VerificationItemReport:
-    """Verify a LinkedIn post by URN presence and format."""
-    post_id = str(post.get("id") or "")
-    ext_id = post.get("external_post_id")
-
-    is_valid_urn = bool(ext_id and ("urn:li:" in ext_id or ext_id.isdigit()))
-    live_url = format_canonical_post_url(platform="linkedin", ext_id=ext_id)
-
-    return VerificationItemReport(
-        post_id=post_id,
-        platform="linkedin",
-        is_verified=is_valid_urn,
-        external_post_id=ext_id,
-        matched_text=post.get("content"),
-        match_confidence=1.0 if is_valid_urn else 0.0,
-        live_url=live_url,
-        status_code=200 if is_valid_urn else 400,
-        error=None if is_valid_urn else "Invalid or missing LinkedIn URN",
-    )
-
-
-def _evaluate_target_post(
-    *, post: dict[str, Any], timeline_tweets: list[dict[str, Any]]
-) -> list[VerificationItemReport]:
-    """Evaluate a single target post across its configured platforms."""
-    platform = str(post.get("platform") or "x").lower().strip()
-    if platform in ("both", "all", "linkx"):
-        return [
-            _verify_single_linkedin_post(post=post),
-            _verify_single_x_post(post=post, timeline_tweets=timeline_tweets),
-        ]
-    if platform == "linkedin":
-        return [_verify_single_linkedin_post(post=post)]
-    return [_verify_single_x_post(post=post, timeline_tweets=timeline_tweets)]
-
-
 async def match_and_verify_posts_node(
     state: VerificationGraphState,
 ) -> dict[str, Any]:
@@ -232,7 +166,7 @@ async def match_and_verify_posts_node(
 
     for post in target_posts:
         post_id = str(post.get("id") or "")
-        reports = _evaluate_target_post(post=post, timeline_tweets=timeline_tweets)
+        reports = evaluate_target_post(post=post, timeline_tweets=timeline_tweets)
         for r in reports:
             items.append(r.model_dump())
